Change data-journey-click strings to boolean on station page * Fix moment.js deprecation warning * Add test for station page
