+++ server/models/views/station.js
@@ -21,12 +21,8 @@ class ViewModel {
     Object.assign(this, {
       feedback: false,
       bingMaps: config.bingKeyMaps,
-      getWarnings: 'Station:Related-content:Get-warnings',
-      planAhead: 'Station:Related-content:Plan-ahead-for-flooding',
-      whatToDo: 'Station:Related-content:What-to-do-in-a-flood',
-      recoverAfter: 'Station:Related-content:Recover-after-a-flood',
-      longTerm: 'Station:Related-content:Check-long-term-risk',
-      reportFlood: 'Station:Related-content:Report-a-flood',
+      getWarnings: true,
+      longTerm: true,
       floodRiskUrl: config.floodRiskUrl,
       trend: river.trend
     })
